Make frame class completely generic; it's just a holder now.
